move non-hmf funcs to lss submodule
