+++ src/halox/__init__.py
@@ -1,6 +1,7 @@
 from . import nfw
 from . import cosmology
 from . import hmf
+from . import lss
 from . import bias
 
-__all__ = ["nfw", "cosmology", "hmf", "bias"]
+__all__ = ["nfw", "cosmology", "hmf", "lss", "bias"]
